+++ scripts/experiments/bertopic_experiment.py
@@ -6,7 +6,7 @@ BERTopic実験スクリプト
 
 import sys
 import os
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '../..'))
 
 import pandas as pd
 from bertopic import BERTopic
