+++ scripts/nlp/extract_topics.py
@@ -6,7 +6,7 @@
 
 import sys
 import os
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '../..'))
 
 import pandas as pd
 from src.nlp.topic import (
